Deletes mixins of executables and modifiables.
