--- src/main/java/spoon/support/delegate/ModifiableDelegateImpl.java
@@ -1,62 +1,0 @@
-package spoon.support.delegate;
-
-import spoon.delegate.ModifiableDelegate;
-import spoon.reflect.declaration.ModifierKind;
-import spoon.support.reflect.declaration.CtElementImpl;
-
-import java.util.Set;
-import java.util.TreeSet;
-
-public class ModifiableDelegateImpl implements ModifiableDelegate {
-	Set<ModifierKind> modifiers = CtElementImpl.EMPTY_SET();
-
-	@Override
-	public Set<ModifierKind> getModifiers() {
-		return modifiers;
-	}
-
-	@Override
-	public boolean hasModifier(ModifierKind modifier) {
-		return getModifiers().contains(modifier);
-	}
-
-	@Override
-	public void setModifiers(Set<ModifierKind> modifiers) {
-		this.modifiers = modifiers;
-	}
-
-	@Override
-	public boolean addModifier(ModifierKind modifier) {
-		if (modifiers == CtElementImpl.<ModifierKind> EMPTY_SET()) {
-			this.modifiers = new TreeSet<ModifierKind>();
-		}
-		return modifiers.add(modifier);
-	}
-
-	@Override
-	public boolean removeModifier(ModifierKind modifier) {
-		return modifiers.remove(modifier);
-	}
-
-	@Override
-	public void setVisibility(ModifierKind visibility) {
-		if (modifiers == CtElementImpl.<ModifierKind> EMPTY_SET()) {
-			this.modifiers = new TreeSet<ModifierKind>();
-		}
-		getModifiers().remove(ModifierKind.PUBLIC);
-		getModifiers().remove(ModifierKind.PROTECTED);
-		getModifiers().remove(ModifierKind.PRIVATE);
-		getModifiers().add(visibility);
-	}
-
-	@Override
-	public ModifierKind getVisibility() {
-		if (getModifiers().contains(ModifierKind.PUBLIC))
-			return ModifierKind.PUBLIC;
-		if (getModifiers().contains(ModifierKind.PROTECTED))
-			return ModifierKind.PROTECTED;
-		if (getModifiers().contains(ModifierKind.PRIVATE))
-			return ModifierKind.PRIVATE;
-		return null;
-	}
-}
