+++ src/main/java/spoon/support/reflect/code/CtLambdaImpl.java
@@ -1,7 +1,6 @@
 package spoon.support.reflect.code;
 
 import spoon.SpoonException;
-import spoon.delegate.ExecutableDelegate;
 import spoon.reflect.code.CtBlock;
 import spoon.reflect.code.CtExpression;
 import spoon.reflect.code.CtLambda;
@@ -9,15 +8,19 @@ import spoon.reflect.declaration.CtParameter;
 import spoon.reflect.reference.CtExecutableReference;
 import spoon.reflect.reference.CtTypeReference;
 import spoon.reflect.visitor.CtVisitor;
-import spoon.support.delegate.ExecutableDelegateImpl;
+import spoon.support.reflect.declaration.CtElementImpl;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+import java.util.TreeSet;
 
 public class CtLambdaImpl<T> extends CtExpressionImpl<T> implements CtLambda<T> {
 	String simpleName;
-	ExecutableDelegate<T> executableDelegate = new ExecutableDelegateImpl<T>();
 	CtExpression<T> expression;
+	CtBlock<?> body;
+	List<CtParameter<?>> parameters = EMPTY_LIST();
+	Set<CtTypeReference<? extends Throwable>> thrownTypes = EMPTY_SET();
 
 	@Override
 	public void accept(CtVisitor visitor) {
@@ -35,8 +38,9 @@ public class CtLambdaImpl<T> extends CtExpressionImpl<T> implements CtLambda<T> 
 	}
 
 	@Override
+	@SuppressWarnings("unchecked")
 	public <B extends T> CtBlock<B> getBody() {
-		return executableDelegate.getBody();
+		return (CtBlock<B>) body;
 	}
 
 	@Override
@@ -44,47 +48,53 @@ public class CtLambdaImpl<T> extends CtExpressionImpl<T> implements CtLambda<T> 
 		if (expression != null) {
 			throw new SpoonException("A lambda can't have two bodys.");
 		}
-		executableDelegate.setBody(body);
+		this.body = body;
 	}
 
 	@Override
 	public List<CtParameter<?>> getParameters() {
-		return executableDelegate.getParameters();
+		return parameters;
 	}
 
 	@Override
 	public void setParameters(List<CtParameter<?>> parameters) {
-		executableDelegate.setParameters(parameters);
+		this.parameters = parameters;
 	}
 
 	@Override
 	public boolean addParameter(CtParameter<?> parameter) {
-		return executableDelegate.addParameter(parameter);
+		if (parameters == CtElementImpl.<CtParameter<?>>EMPTY_LIST()) {
+			parameters = new ArrayList<CtParameter<?>>();
+		}
+		return parameters.add(parameter);
 	}
 
 	@Override
 	public boolean removeParameter(CtParameter<?> parameter) {
-		return executableDelegate.removeParameter(parameter);
+		return parameters.remove(parameter);
 	}
 
 	@Override
 	public Set<CtTypeReference<? extends Throwable>> getThrownTypes() {
-		return executableDelegate.getThrownTypes();
+		return thrownTypes;
 	}
 
 	@Override
 	public void setThrownTypes(Set<CtTypeReference<? extends Throwable>> thrownTypes) {
-		executableDelegate.setThrownTypes(thrownTypes);
+		this.thrownTypes = thrownTypes;
 	}
 
 	@Override
 	public boolean addThrownType(CtTypeReference<? extends Throwable> throwType) {
-		return executableDelegate.addThrownType(throwType);
+		if (thrownTypes == CtElementImpl.<CtTypeReference<? extends Throwable>>EMPTY_SET()) {
+			thrownTypes = new TreeSet<CtTypeReference<? extends Throwable>>();
+		}
+		return thrownTypes.add(throwType);
 	}
 
 	@Override
 	public boolean removeThrownType(CtTypeReference<? extends Throwable> throwType) {
-		return executableDelegate.removeThrownType(throwType);
+		return thrownTypes.remove(throwType);
 	}
 
 	@Override
@@ -99,7 +109,7 @@ public class CtLambdaImpl<T> extends CtExpressionImpl<T> implements CtLambda<T> 
 
 	@Override
 	public void setExpression(CtExpression<T> expression) {
-		if (executableDelegate.getBody() != null) {
+		if (body != null) {
 			throw new SpoonException("A lambda can't have two bodys.");
 		}
 		this.expression = expression;
