+++ src/main/java/spoon/support/reflect/declaration/CtExecutableImpl.java
@@ -46,54 +46,61 @@ public abstract class CtExecutableImpl<R> extends CtNamedElementImpl implements 
 		super();
 	}
 
-	public boolean addParameter(CtParameter<?> parameter) {
-		if (parameters == CtElementImpl.<CtParameter<?>> EMPTY_LIST()) {
-			parameters = new ArrayList<CtParameter<?>>();
-		}
-		return parameters.add(parameter);
-	}
-
-	public boolean removeParameter(CtParameter<?> parameter) {
-		return parameters.remove(parameter);
-	}
-
-	public boolean addThrownType(CtTypeReference<? extends Throwable> throwType) {
-		if (thrownTypes == CtElementImpl
-				.<CtTypeReference<? extends Throwable>> EMPTY_SET()) {
-			thrownTypes = new TreeSet<CtTypeReference<? extends Throwable>>();
-		}
-		return thrownTypes.add(throwType);
-	}
-
-	public boolean removeThrownType(
-			CtTypeReference<? extends Throwable> throwType) {
-		return thrownTypes.remove(throwType);
-	}
-
+	@Override
 	@SuppressWarnings("unchecked")
 	public <B extends R> CtBlock<B> getBody() {
 		return (CtBlock<B>) body;
 	}
 
-	public List<CtParameter<?>> getParameters() {
-		return parameters;
-	}
-
-	public Set<CtTypeReference<? extends Throwable>> getThrownTypes() {
-		return thrownTypes;
-	}
-
+	@Override
 	public <B extends R> void setBody(CtBlock<B> body) {
 		this.body = body;
 	}
 
+	@Override
+	public List<CtParameter<?>> getParameters() {
+		return parameters;
+	}
+
+	@Override
 	public void setParameters(List<CtParameter<?>> parameters) {
 		this.parameters = parameters;
 	}
 
-	public void setThrownTypes(
-			Set<CtTypeReference<? extends Throwable>> thrownTypes) {
+	@Override
+	public boolean addParameter(CtParameter<?> parameter) {
+		if (parameters == CtElementImpl.<CtParameter<?>>EMPTY_LIST()) {
+			parameters = new ArrayList<CtParameter<?>>();
+		}
+		return parameters.add(parameter);
+	}
+
+	@Override
+	public boolean removeParameter(CtParameter<?> parameter) {
+		return parameters.remove(parameter);
+	}
+
+	@Override
+	public Set<CtTypeReference<? extends Throwable>> getThrownTypes() {
+		return thrownTypes;
+	}
+
+	@Override
+	public void setThrownTypes(Set<CtTypeReference<? extends Throwable>> thrownTypes) {
 		this.thrownTypes = thrownTypes;
+	}
+
+	@Override
+	public boolean addThrownType(CtTypeReference<? extends Throwable> throwType) {
+		if (thrownTypes == CtElementImpl.<CtTypeReference<? extends Throwable>>EMPTY_SET()) {
+			thrownTypes = new TreeSet<CtTypeReference<? extends Throwable>>();
+		}
+		return thrownTypes.add(throwType);
+	}
+
+	@Override
+	public boolean removeThrownType(CtTypeReference<? extends Throwable> throwType) {
+		return thrownTypes.remove(throwType);
 	}
 
 	@Override
